requirements.txt
aliyun-python-sdk-core
aliyun-python-sdk-cms
python-docx
pandas
matplotlib